Convert some utility functions to extensions
