+++ packages/winrtgen/lib/src/projections/winrt_enum.dart
@@ -4,6 +4,7 @@
 
 import 'package:winmd/winmd.dart';
 
+import '../extensions.dart';
 import '../utils.dart';
 
 /// Represents a Dart projection of a WinRT enumeration typedef.
