+++ packages/winrtgen/lib/src/projections/winrt_property.dart
@@ -2,6 +2,7 @@
 // details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import '../extensions.dart';
 import '../utils.dart';
 import 'winrt_method.dart';
